Consistently use magma_int_t. Also delete unnecesary static, to make (closer to) thread safe.
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@2057 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ include/magmablas.h
@@ -12,8 +12,8 @@
 #include <cublas.h>
 #include <cuda.h>
 
-typedef int magma_int_t;
-typedef int magma_err_t;
+typedef long magma_int_t;
+typedef int  magma_err_t;
 typedef void* magma_devptr;
 
 // For now, make these compatible with old cublas v1 prototypes.
