+++ magmablas/chemv_tesla.cu
@@ -941,8 +941,8 @@ magmablas_chemv_130( char uplo, magma_int_t n,
                      cuFloatComplex beta,  
                      cuFloatComplex *Y, magma_int_t incy)
 {
-    char      uplo_[2] = {uplo, 0};
-    long int  upper    = lapackf77_lsame(uplo_, "U");
+    char uplo_[2] = {uplo, 0};
+    int  upper    = lapackf77_lsame(uplo_, "U");
 
     /*
      * Test the input parameters.
