+++ magmablas/zhemv_fermi_mgpu_offset.cu
@@ -1023,8 +1023,8 @@ magmablas_zhemv_200_mgpu_offset( char uplo, magma_int_t n,
               cudaStream_t stream[][10])
 
 {
-    char      uplo_[2] = {uplo, 0};
-    long int  upper    = lapackf77_lsame(uplo_, "U");
+    char uplo_[2] = {uplo, 0};
+    int  upper    = lapackf77_lsame(uplo_, "U");
 
 
 
@@ -1123,8 +1123,8 @@ magmablas_zhemv2_200_mgpu_offset( char uplo, magma_int_t n,
               magma_int_t offset)
 
 {
-    char      uplo_[2] = {uplo, 0};
-    long int  upper    = lapackf77_lsame(uplo_, "U");
+    char uplo_[2] = {uplo, 0};
+    int  upper    = lapackf77_lsame(uplo_, "U");
