+++ magmablas/zhemv_tesla.cu
@@ -200,8 +200,8 @@ magmablas_zhemv_130( char uplo, magma_int_t n,
                        cuDoubleComplex beta,  
                        cuDoubleComplex *Y, magma_int_t incy)
 {
-    char      uplo_[2] = {uplo, 0};
-    long int  upper    = lapackf77_lsame(uplo_, "U");
+    char uplo_[2] = {uplo, 0};
+    int  upper    = lapackf77_lsame(uplo_, "U");
 
     /*
      * Test the input parameters.
