+++ src/dbulge_aux.cpp
@@ -49,7 +49,7 @@ extern "C" void  magma_dstedc_withZ(char JOBZ, magma_int_t N, double *D, double 
 
   if(INFO!=0){
         printf("=================================================\n");
-        printf("DSTEDC ERROR OCCURED. HERE IS INFO %d \n ",INFO);
+        printf("DSTEDC ERROR OCCURED. HERE IS INFO %d \n ", (int) INFO);
         printf("=================================================\n");
           //assert(INFO==0);
   }
@@ -88,7 +88,7 @@ extern "C" void  magma_dstedx_withZ(magma_int_t N, magma_int_t NE, double *D, do
 
   if(INFO!=0){
         printf("=================================================\n");
-        printf("DSTEDC ERROR OCCURED. HERE IS INFO %d \n ",INFO);
+        printf("DSTEDC ERROR OCCURED. HERE IS INFO %d \n ", (int) INFO);
         printf("=================================================\n");
           //assert(INFO==0);
   }
