+++ src/zgeqrf-v2.cpp
@@ -168,7 +168,7 @@ magma_zgeqrf2(magma_context *cntxt, magma_int_t m, magma_int_t n,
 
     int lwkopt = n * nb;
     work[0] = MAGMA_Z_MAKE( (double)lwkopt, 0 );
-    long int lquery = (lwork == -1);
+    int lquery = (lwork == -1);
     if (m < 0) {
         *info = -1;
     } else if (n < 0) {
