+++ testing/testing_zcgesv_gpu.cpp
@@ -72,7 +72,7 @@ int main(int argc , char **argv)
     }
     else {
         printf("\nUsage: \n");
-        printf("  testing_zcgesv_gpu -nrhs %d -N %d\n\n", NRHS, 1024);
+        printf("  testing_zcgesv_gpu -nrhs %d -N %d\n\n", (int) NRHS, 1024);
     }
     printf("Epsilon(double): %8.6e\n"
            "Epsilon(single): %8.6e\n\n", 
@@ -115,7 +115,7 @@ int main(int argc , char **argv)
         lapackf77_zlarnv( &ione, ISEED, &size, h_B );
         lapackf77_zlacpy( MagmaUpperLowerStr, &N, &NRHS, h_B, &ldb, h_X, &ldx);
 
-        printf("%5d  ",N);
+        printf("%5d  ",(int) N);
 
         magma_zsetmatrix( N, N,    h_A, lda, d_A, ldda );
         magma_zsetmatrix( N, NRHS, h_B, ldb, d_B, lddb );
@@ -130,7 +130,7 @@ int main(int argc , char **argv)
                           d_WORKD, d_WORKS, &iter, &info);
         end = get_current_time();
         if (info < 0)
-            printf("Argument %d of magma_zcgesv had an illegal value.\n", -info);
+            printf("Argument %d of magma_zcgesv had an illegal value.\n", (int) -info);
         gpu_perf = flopsS / GetTimerValue(start, end);
 
         //=====================================================================
@@ -155,7 +155,7 @@ int main(int argc , char **argv)
         magma_zgetrf_gpu(N, N, d_A, ldda, h_ipiv, &info);
         end = get_current_time();
         if (info < 0)
-            printf("Argument %d of magma_zgetrf had an illegal value.\n", -info);
+            printf("Argument %d of magma_zgetrf had an illegal value.\n", (int) -info);
         gpu_perfdf = flopsF / GetTimerValue(start, end);
 
         printf("%6.2f    ", gpu_perfdf); fflush(stdout);
@@ -171,7 +171,7 @@ int main(int argc , char **argv)
         magma_zgetrs_gpu( trans, N, NRHS, d_A, ldda, h_ipiv, d_B, lddb, &info );
         end = get_current_time();
         if (info < 0)
-            printf("Argument %d of magma_zgetrs had an illegal value.\n", -info);
+            printf("Argument %d of magma_zgetrs had an illegal value.\n", (int) -info);
 
         gpu_perfds = flopsS / GetTimerValue(start, end);
 
@@ -191,7 +191,7 @@ int main(int argc , char **argv)
         magma_cgetrf_gpu(N, N, d_As, ldda, h_ipiv, &info);
         end = get_current_time();
         if (info < 0)
-            printf("Argument %d of magma_cgetrf had an illegal value.\n", -info);
+            printf("Argument %d of magma_cgetrf had an illegal value.\n", (int) -info);
         
         gpu_perfsf = flopsF / GetTimerValue(start, end);
         printf("%6.2f     ", gpu_perfsf); fflush(stdout);
@@ -208,13 +208,13 @@ int main(int argc , char **argv)
                           d_Bs, lddb, &info);
         end = get_current_time();
         if (info < 0)
-            printf("Argument %d of magma_cgetrs had an illegal value.\n", -info);
+            printf("Argument %d of magma_cgetrs had an illegal value.\n", (int) -info);
 
         gpu_perfss = flopsS / GetTimerValue(start, end);
         printf("%6.2f     ", gpu_perfss); fflush(stdout);
 
         printf("%6.2f     ", gpu_perf);
-        printf("%e    %3d\n", Rnorm/Anorm, iter); fflush(stdout);
+        printf("%e    %3d\n", Rnorm/Anorm, (int) iter); fflush(stdout);
 
         if( argc != 1 ) break ;        
     }
